+++ examples/bufferless-slab.rs
@@ -3,8 +3,9 @@
 use async_net::TcpListener;
 use futures_util::StreamExt;
 use ruchei::{
-    concurrent::ConcurrentExt, echo::bufferless::EchoBufferless, multi_item::MultiItemExt,
-    multicast::bufferless::MulticastBufferless, poll_on_wake::PollOnWakeExt,
+    concurrent::ConcurrentExt, connection_item::ConnectionItemExt,
+    echo::bufferless::EchoBufferless, multicast::bufferless::MulticastBufferless,
+    poll_on_wake::PollOnWakeExt,
 };
 
 #[async_std::main]
@@ -20,7 +21,7 @@ async fn main() {
         .concurrent()
         .filter_map(|r| async { r.ok() })
         .multicast_bufferless()
-        .multi_item_ignore()
+        .connection_item_ignore()
         .echo_bufferless()
         .await
         .unwrap();