+++ examples/bufferless-trie.rs
@@ -4,8 +4,8 @@ use async_net::TcpListener;
 use futures_util::{SinkExt, StreamExt, TryStreamExt, future::ready};
 use ruchei::{
     concurrent::ConcurrentExt,
+    connection_item::ConnectionItemExt,
     echo::bufferless::EchoBufferless,
-    multi_item::MultiItemExt,
     multicast::trie::{MulticastTrie, SubRequest},
     poll_on_wake::PollOnWakeExt,
 };
@@ -31,7 +31,7 @@ async fn main() {
         })
         .map(|s| s.with(|(_, m)| ready(Ok::<_, async_tungstenite::tungstenite::Error>(m))))
         .multicast_trie()
-        .multi_item_ignore()
+        .connection_item_ignore()
         .map_ok(|m| ("prefix", m))
         .echo_bufferless()
         .await