+++ examples/echo-dealer-slab.rs
@@ -3,8 +3,9 @@
 use async_net::TcpListener;
 use futures_util::{StreamExt, future::ready};
 use ruchei::{
-    concurrent::ConcurrentExt, deal::without_multicast::DealWithoutMulticast,
-    echo::bufferless::EchoBufferless, multi_item::MultiItemExt, poll_on_wake::PollOnWakeExt,
+    concurrent::ConcurrentExt, connection_item::ConnectionItemExt,
+    deal::without_multicast::DealWithoutMulticast, echo::bufferless::EchoBufferless,
+    poll_on_wake::PollOnWakeExt,
 };
 
 #[async_std::main]
@@ -21,7 +22,7 @@ async fn main() {
         .filter_map(|r| async { r.ok() })
         .map(|s| s.filter(|m| ready(m.as_ref().is_ok_and(|m| !m.is_close()))))
         .deal_without_multicast()
-        .multi_item_ignore()
+        .connection_item_ignore()
         .echo_bufferless()
         .await
         .unwrap();