+++ examples/echo-router-keyed.rs
@@ -3,7 +3,7 @@
 use async_net::TcpListener;
 use futures_util::StreamExt;
 use ruchei::{
-    concurrent::ConcurrentExt, echo::route::EchoRoute, multi_item::MultiItemExt,
+    concurrent::ConcurrentExt, connection_item::ConnectionItemExt, echo::route::EchoRoute,
     poll_on_wake::PollOnWakeExt, route::keyed::RouterKeyedExt,
 };
 
@@ -21,7 +21,7 @@ async fn main() {
         .filter_map(|r| async { r.ok() })
         .map(|s| (rand::random::<u64>(), s))
         .route_keyed()
-        .multi_item_ignore()
+        .connection_item_ignore()
         .echo_route()
         .await
         .unwrap();