+++ src/deal/without_multicast.rs
@@ -18,7 +18,7 @@ use ruchei_collections::{
 use ruchei_connection::{Connection, ConnectionWaker, Ready};
 use ruchei_extend::{Extending, ExtendingExt};
 
-use crate::multi_item::MultiItem;
+use crate::connection_item::ConnectionItem;
 
 const OP_WAKE_NEXT: usize = 0;
 const OP_DEAL: usize = 1;
@@ -169,14 +169,14 @@ impl<E, Out, S: Unpin + Sink<Out, Error = E>> Dealer<S, Out, E> {
 impl<In, Out, E, S: Unpin + TryStream<Ok = In, Error = E> + Sink<Out, Error = E>> Stream
     for Dealer<S, Out, E>
 {
-    type Item = MultiItem<S>;
+    type Item = ConnectionItem<S>;
 
     fn poll_next(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Option<Self::Item>> {
         self.wakers.next.register(cx.waker());
         let _ = self.as_mut().poll();
         let mut this = self.as_mut().project();
         if let Some((stream, error)) = this.closed.pop_front() {
-            return Poll::Ready(Some(MultiItem::Closed(stream, error)));
+            return Poll::Ready(Some(ConnectionItem::Closed(stream, error)));
         }
         this.next.register(cx);
         while let Some(key) = this.next.as_mut().next::<OP_WAKE_NEXT>(this.connections) {
@@ -188,7 +188,7 @@ impl<In, Out, E, S: Unpin + TryStream<Ok = In, Error = E> + Sink<Out, Error = E>
                 match o {
                     Some(Ok(item)) => {
                         this.next.downgrade().insert(key);
-                        return Poll::Ready(Some(MultiItem::Item(item)));
+                        return Poll::Ready(Some(ConnectionItem::Item(item)));
                     }
                     Some(Err(e)) => {
                         self.as_mut().remove(key, Some(e));