+++ src/liveness/group_concurrent.rs
@@ -15,9 +15,9 @@
 //! # use futures_util::{future::ready, StreamExt};
 //! # use ruchei::{
 //! #     concurrent::ConcurrentExt,
+//! #     connection_item::ConnectionItemExt,
 //! #     echo::buffered::EchoBuffered,
 //! #     liveness::{group_concurrent::{Group, GroupConcurrent}, timeout_unused::TimeoutUnused},
-//! #     multi_item::MultiItemExt,
 //! #     multicast::replay::MulticastReplay,
 //! #     poll_on_wake::PollOnWakeExt,
 //! # };
@@ -68,7 +68,7 @@
 //!                 .timeout_unused(|| ready(()))
 //!                 // All messages within a group are replayed on new connection joining
 //!                 .multicast_replay()
-//!                 .multi_item_ignore()
+//!                 .connection_item_ignore()
 //!                 .echo_buffered()
 //!                 .await
 //!                 .unwrap();