+++ src/multicast/trie.rs
@@ -17,7 +17,7 @@ use ruchei_collections::{
 use ruchei_connection::{Connection, ConnectionWaker, Ready};
 use ruchei_extend::{Extending, ExtendingExt};
 
-use crate::multi_item::MultiItem;
+use crate::connection_item::ConnectionItem;
 
 const OP_WAKE_NEXT: usize = 0;
 const OP_WAKE_READY: usize = 1;
@@ -102,12 +102,12 @@ impl<S, E> Multicast<S, E> {
 impl<K: AsRef<[u8]>, O, E, S: Unpin + TryStream<Ok = SubRequest<K, O>, Error = E>> Stream
     for Multicast<S, E>
 {
-    type Item = MultiItem<S, O>;
+    type Item = ConnectionItem<S, O>;
 
     fn poll_next(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Option<Self::Item>> {
         let mut this = self.as_mut().project();
         if let Some((stream, error)) = this.closed.pop_front() {
-            return Poll::Ready(Some(MultiItem::Closed(stream, error)));
+            return Poll::Ready(Some(ConnectionItem::Closed(stream, error)));
         }
         this.next.register(cx);
         while let Some(key) = this.next.as_mut().next::<OP_WAKE_NEXT>(this.connections) {
@@ -133,7 +133,7 @@ impl<K: AsRef<[u8]>, O, E, S: Unpin + TryStream<Ok = SubRequest<K, O>, Error = E
                                 this.connections.mt_remove(&key, route.as_ref())
                             }
                             SubRequest::Other(item) => {
-                                return Poll::Ready(Some(MultiItem::Item(item)));
+                                return Poll::Ready(Some(ConnectionItem::Item(item)));
                             }
                         }
                     }