+++ src/route/multicast.rs
@@ -16,7 +16,7 @@ use ruchei_collections::{
 use ruchei_connection::{Connection, ConnectionWaker, Ready};
 use ruchei_extend::{Extending, ExtendingExt};
 
-use crate::multi_item::{MultiItem, MultiRouteItem};
+use crate::connection_item::{ConnectionItem, MultiRouteItem};
 
 const OP_WAKE_NEXT: usize = 0;
 const OP_WAKE_READY: usize = 1;
@@ -196,7 +196,7 @@ impl<In, E, S: Unpin + TryStream<Ok = In, Error = E>> Stream for Router<S, E> {
     fn poll_next(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Option<Self::Item>> {
         let mut this = self.as_mut().project();
         if let Some((key, stream, error)) = this.closed.pop_front() {
-            return Poll::Ready(Some(MultiItem::Closed((key, stream), error)));
+            return Poll::Ready(Some(ConnectionItem::Closed((key, stream), error)));
         }
         this.next.register(cx);
         while let Some(key) = this.next.as_mut().next::<OP_WAKE_NEXT>(this.connections) {
@@ -211,7 +211,7 @@ impl<In, E, S: Unpin + TryStream<Ok = In, Error = E>> Stream for Router<S, E> {
                 match o {
                     Some(Ok(item)) => {
                         this.next.downgrade().insert(key);
-                        return Poll::Ready(Some(MultiItem::Item((RouteKey(key), item))));
+                        return Poll::Ready(Some(ConnectionItem::Item((RouteKey(key), item))));
                     }
                     Some(Err(e)) => {
                         self.as_mut().remove(key, Some(e));